+++ original/Code/Combat/diaglog.cpp
@@ -66,6 +66,6 @@ void	DiagLogClass::Log_Timed( const char * type, const char * format, ... )
 		StringClass line;
 		float time = TimeManager::Get_Total_Seconds();
 		line.Format( "%s; %1.2f; %s%c%c", type, time, (const char*)data, 0x0D, 0x0A );
-		_DiagLogFile->Write( line, ::strlen( line ) );
+		_DiagLogFile->Write( line, static_cast<int32_t>(::strlen( line )) );
 	}
 }
